requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.4.26
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tornado==6.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yt-dlp==2025.3.31
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.4.26
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tornado==6.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yt-dlp==2025.3.31
nest_asyncio==1.6.0

